refactor(installations): split modular and namespaced TypeScript surfaces
