+++ packages/installations/lib/index.ts
@@ -15,149 +15,13 @@
  *
  */
 
-import { isIOS } from '@react-native-firebase/app/dist/module/common';
-import type { ReactNativeFirebase } from '@react-native-firebase/app';
-import {
-  createModuleNamespace,
-  FirebaseModule,
-  getFirebaseRoot,
-} from '@react-native-firebase/app/dist/module/internal';
+// Export modular/public types.
+export type * from './types/installations';
 
-import { version } from './version';
-
-const statics = {};
-
-const namespace = 'installations';
-
-const nativeModuleName = 'RNFBInstallationsModule';
-
-interface InstallationsNativeModule {
-  getId(): Promise<string>;
-  getToken(forceRefresh: boolean): Promise<string>;
-  delete(): Promise<void>;
-}
-
-/**
- * Firebase Installations package for React Native.
- *
- * @firebase installations
- */
-export namespace FirebaseInstallationsTypes {
-  export interface Statics {
-    SDK_VERSION: string;
-  }
-
-  /**
-   * The Firebase Installations service is available for the default app or a given app.
-   */
-  export interface Module extends ReactNativeFirebase.FirebaseModule {
-    /**
-     * The current `FirebaseApp` instance for this Firebase service.
-     */
-    app: ReactNativeFirebase.FirebaseApp;
-
-    /**
-     * Creates a Firebase Installation if there isn't one for the app and returns the Installation ID.
-     */
-    getId(): Promise<string>;
-
-    /**
-     * Retrieves a valid installation auth token.
-     */
-    getToken(forceRefresh?: boolean): Promise<string>;
-
-    /**
-     * Deletes the Firebase Installation and all associated data from the Firebase backend.
-     */
-    delete(): Promise<void>;
-  }
-}
-
-class FirebaseInstallationsModule extends FirebaseModule {
-  get native(): InstallationsNativeModule {
-    return super.native as unknown as InstallationsNativeModule;
-  }
-
-  getId(): Promise<string> {
-    return this.native.getId();
-  }
-
-  getToken(forceRefresh?: boolean): Promise<string> {
-    if (!forceRefresh) {
-      return this.native.getToken(false);
-    } else {
-      return this.native.getToken(true);
-    }
-  }
-
-  delete(): Promise<void> {
-    return this.native.delete();
-  }
-
-  onIdChange(): () => void {
-    if (isIOS) {
-      return () => {};
-    }
-
-    // TODO implement change listener on Android
-    return () => {};
-  }
-}
-
-type InstallationsNamespace = ReactNativeFirebase.FirebaseModuleWithStaticsAndApp<
-  FirebaseInstallationsTypes.Module,
-  FirebaseInstallationsTypes.Statics
-> & {
-  firebase: ReactNativeFirebase.Module;
-  app(name?: string): ReactNativeFirebase.FirebaseApp;
-};
-
-// import { SDK_VERSION } from '@react-native-firebase/installations';
-export const SDK_VERSION = version;
-
-// import installations from '@react-native-firebase/installations';
-// installations().X(...);
-const defaultExport = createModuleNamespace({
-  statics,
-  version,
-  namespace,
-  nativeModuleName,
-  nativeEvents: false, // TODO implement android id change listener: ['installations_id_changed'],
-  hasMultiAppSupport: true,
-  hasCustomUrlOrRegionSupport: false,
-  ModuleClass: FirebaseInstallationsModule,
-}) as unknown as InstallationsNamespace;
-
-export default defaultExport;
-
-// import installations, { firebase } from '@react-native-firebase/installations';
-// installations().X(...);
-// firebase.installations().X(...);
-export const firebase = getFirebaseRoot() as unknown as ReactNativeFirebase.Module & {
-  installations: typeof defaultExport;
-  app(
-    name?: string,
-  ): ReactNativeFirebase.FirebaseApp & { installations(): FirebaseInstallationsTypes.Module };
-};
-
+// Export modular API functions.
 export * from './modular';
 
-/**
- * Attach namespace to `firebase.` and `FirebaseApp.`.
- */
-declare module '@react-native-firebase/app' {
-  namespace ReactNativeFirebase {
-    import FirebaseModuleWithStaticsAndApp = ReactNativeFirebase.FirebaseModuleWithStaticsAndApp;
-
-    interface Module {
-      installations: FirebaseModuleWithStaticsAndApp<
-        FirebaseInstallationsTypes.Module,
-        FirebaseInstallationsTypes.Statics
-      >;
-    }
-
-    interface FirebaseApp {
-      installations(): FirebaseInstallationsTypes.Module;
-    }
-  }
-}
+// Export namespaced API.
+export type { FirebaseInstallationsTypes } from './types/namespaced';
+export * from './namespaced';
+export { default } from './namespaced';
