+++ packages/installations/lib/types/installations.ts
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2016-present Invertase Limited & Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this library except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import type { FirebaseApp, ReactNativeFirebase } from '@react-native-firebase/app';
+
+/**
+ * Firebase Installations service instance.
+ */
+export interface Installations extends ReactNativeFirebase.FirebaseModule {
+  /**
+   * The current `FirebaseApp` instance for this Firebase service.
+   */
+  app: FirebaseApp;
+
+  /**
+   * Creates a Firebase Installation if there isn't one for the app and returns the Installation ID.
+   */
+  getId(): Promise<string>;
+
+  /**
+   * Retrieves a valid installation auth token.
+   */
+  getToken(forceRefresh?: boolean): Promise<string>;
+
+  /**
+   * Deletes the Firebase Installation and all associated data from the Firebase backend.
+   */
+  delete(): Promise<void>;
+}
